[bridge 17/n] add real abis
## Description 

This PR adds real ABIs generated from solidity code

## Test Plan 

unit tests and integration tests in stacked PRs

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes

+++ crates/haneul-bridge/src/events.rs
@@ -22,7 +22,7 @@ use haneul_json_rpc_types::HaneulEvent;
 use haneul_types::base_types::HaneulAddress;
 use haneul_types::digests::TransactionDigest;
 
-// TODO: Placeholder, this will need to match the actual event types defined in Move
+// This is the event structure defined and emitted in Move
 #[derive(Debug, Serialize, Deserialize, PartialEq, Eq, Clone)]
 pub struct EmittedHaneulToEthTokenBridgeV1 {
     pub nonce: u64,
@@ -31,7 +31,7 @@ pub struct EmittedHaneulToEthTokenBridgeV1 {
     pub haneul_address: HaneulAddress,
     pub eth_address: EthAddress,
     pub token_id: TokenId,
-    pub amount: u128,
+    pub amount: u64,
 }
 
 const EMITTED_HANEUL_TO_ETH_TOKEN_BRIDGE_V1_STUCT_TAG: &str =
